Add files via upload
This is my Simple project about medicine purchasing in shop. This project is made with Java language for my university OOP1 course. Me and my team just used some simple Java principles in this project. And also makes it interactive with Java GUI(Graphical user interface). Click the photo or title for the GitHub link.